update waitEventCount int to size_t for consistency

+++ include/ThreadPool.h
@@ -36,7 +36,7 @@ private:
     std::atomic<size_t> totalTasksExecuted{0};
     std::atomic<size_t> totalExecutionTime{0};
     std::atomic<size_t> totalWaitingTime{0};
-    std::atomic<int> waitEventCount{0};
+    std::atomic<size_t> waitEventCount{0};
     std::atomic<size_t> totalQueueLength{0};
     std::atomic<size_t> queueCheckCount{0};
     void printStats() const;
